refactor: split shared model package
